feat: 使用markdown渲染题干和选项 feat: 接口调整

+++ lib/model/user_info.dart
@@ -26,7 +26,7 @@ class UserInfo {
   int peeProfession;
 
   /// b币
-  int bCoin;
+  double bCoin;
 
   /// 金瓜子
   int gSeed;
